+++ src/sidepanel/pages/TaskBoard/index.jsx
@@ -1,55 +1,36 @@
 import "@/styles/styles.css";
 import { useEffect, useState, useRef } from "react";
-import { useNavigate } from "react-router-dom";
 import { v4 as uuidv4 } from "uuid";
 
-import createManual from "@/api/createManual";
+import useTaskHandlers from "@/hooks/useTaskHandlers";
 import LoadingModal from "@/sidepanel/components/LoadingModal";
 import WarningModal from "@/sidepanel/components/WarningModal";
-import { getCaptureStatus, resetCapturedSteps } from "@/utils/storage";
 
 import TaskControlBar from "./TaskControlBar";
 import TaskStatusHeader from "./TaskStatusHeader";
 import TaskStepsRenderer from "./TaskStepsRenderer";
 
 const TaskBoard = () => {
-  const navigate = useNavigate();
-
   const [steps, setSteps] = useState([]);
   const [isCapturing, setIsCapturing] = useState(true);
   const [isLoading, setIsLoading] = useState(false);
   const [showModal, setShowModal] = useState(false);
 
+  const {
+    handleDeleteStep,
+    handleTitleChange,
+    handlePauseClick,
+    handleFinishClick,
+    handleCleanupClick,
+  } = useTaskHandlers({
+    steps,
+    setSteps,
+    setIsCapturing,
+    setIsLoading,
+    setShowModal,
+  });
+
   const isCapturingRef = useRef(isCapturing);
-
-  const goBack = () => {
-    navigate("/");
-  };
-  const handleTitleChange = (index, newTitle) => {
-    setSteps((prev) => {
-      const updated = [...prev];
-      updated[index].elementData.textContent = newTitle;
-
-      chrome.storage.local.set({ CapturedSteps: updated });
-      return updated;
-    });
-  };
-  const handleCleanupClick = async () => {
-    chrome.runtime.sendMessage({ type: "CLEANUP_ALL" }, async () => {
-      if (chrome.runtime.lastError) {
-        console.error("CLEANUP_ALL 전송 오류:", chrome.runtime.lastError.message);
-      } else {
-        console.log("📨 CLEANUP_ALL 메시지 전송됨");
-        await chrome.storage.local.set({ isCapturing: false });
-
-        const status = await getCaptureStatus();
-        setIsCapturing(status);
-
-        goBack();
-      }
-    });
-    resetCapturedSteps();
-  };
 
   useEffect(() => {
     chrome.storage.local.get("CapturedSteps").then(({ CapturedSteps }) => {
@@ -57,12 +38,7 @@ const TaskBoard = () => {
         setSteps(CapturedSteps);
       }
     });
-    chrome.runtime.sendMessage({ type: "START_CAPTURE" }, () => {
-      if (chrome.runtime.lastError) {
-        console.error("START_CAPTURE 에러:", chrome.runtime.lastError);
-        return;
-      }
-    });
+    chrome.runtime.sendMessage({ type: "START_CAPTURE" });
 
     const handleStorageChange = (changes, areaName) => {
       if (areaName === "local" && changes.CapturedSteps && isCapturingRef.current) {
@@ -92,43 +68,6 @@ const TaskBoard = () => {
   useEffect(() => {
     isCapturingRef.current = isCapturing;
   }, [isCapturing]);
-
-  const handleDeleteStep = (taskId) => {
-    setSteps((prev) => {
-      const updatedSteps = prev.filter((step) => step.id !== taskId);
-      chrome.storage.local.set({ CapturedSteps: updatedSteps });
-      return updatedSteps;
-    });
-  };
-
-  const handlePauseClick = () => {
-    setIsCapturing((prev) => {
-      chrome.storage.local.set({ isCapturing: !prev });
-      return !prev;
-    });
-  };
-
-  const handleFinishClick = async () => {
-    const stepData = steps.map((step) => ({
-      text: step.elementData.textContent,
-      image: step.image,
-    }));
-    const body = {
-      steps: stepData,
-    };
-
-    setIsLoading(true);
-    try {
-      await createManual(body);
-      resetCapturedSteps();
-      navigate("/repository");
-    } catch (error) {
-      console.error("매뉴얼 생성 에러:", error);
-      setShowModal(true);
-    } finally {
-      setIsLoading(false);
-    }
-  };
 
   return (
     <div className="flex min-h-screen w-full flex-col bg-white">
